fix: Top should allow uparrown/downarrow to cycle through namespaces
also refactors top controller into its own subdir, and refactors it into smaller files in that dir
