+++ plugins/plugin-codeflare-dashboard/src/controller/dashboard/top/index.ts
@@ -1,0 +1,55 @@
+/*
+ * Copyright 2023 The Kubernetes Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import Debug from "debug"
+
+import type { Arguments } from "@kui-shell/core"
+import type TopOptions from "./options.js"
+
+import { enterAltBufferMode } from "../term.js"
+import { getCurrentCluster, getCurrentNamespace, changeContext } from "../../kubernetes.js"
+
+import initWatcher from "./watcher.js"
+
+export async function getNamespaceFromArgsOrCurrent(args: Arguments<TopOptions>) {
+  return /*args.parsedOptions.A ? "All Namespaces" :*/ args.parsedOptions.n || (await getCurrentNamespace())
+}
+
+export default async function jobsController(args: Arguments<TopOptions>) {
+  const debug = Debug("plugin-codeflare-dashboard/controller/top")
+
+  if (process.env.ALT !== "false") {
+    enterAltBufferMode()
+  }
+
+  // these will be the initial values of cluster and namespace focus
+  const [cluster, ns] = await Promise.all([getCurrentCluster(), getNamespaceFromArgsOrCurrent(args)])
+  debug("cluster", cluster)
+  debug("namespace", ns || "using namespace from user current context")
+
+  debug("loading UI dependencies")
+  const [{ default: render }] = await Promise.all([import("../../../components/Top/index.js")])
+
+  debug("rendering")
+  await render({
+    cluster,
+    namespace: ns,
+    initWatcher: initWatcher.bind(args.parsedOptions),
+    changeContext,
+  })
+  debug("exiting")
+  return true
+}
